refactor model folder
make model folder
move gitignore
