--- .gitignore
@@ -1,5 +1,0 @@
-/junbae/saved/
-/junbae/submission/
-/junbae/config.py
-/junbae/wandb/
-/junbae/__pycache__/
