+++ junbae/models/FCN32s.py
@@ -1,0 +1,103 @@
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+from torchvision import models
+from torchvision.models import vgg16
+import segmentation_models_pytorch as smp
+from utils import seed_everything
+
+seed_everything(21)
+
+
+class FCN32s(nn.Module):
+    def __init__(self, num_classes=12):
+        super(FCN32s, self).__init__()
+
+        def CBR(in_channels, out_channels, kernel_size=3, stride=1, padding=1):
+            # convolution
+            # relu
+            # 한블럭
+            return nn.Sequential(
+                nn.Conv2d(in_channels=in_channels, out_channels=out_channels,
+                          kernel_size=kernel_size, stride=stride, padding=padding),
+                nn.ReLU(inplace=True))
+
+        # conv1
+        self.conv1_1 = CBR(3, 64, 3, 1, 1)
+        self.conv1_2 = CBR(64, 64, 3, 1, 1)
+        self.pool1 = nn.MaxPool2d(2, stride=2, ceil_mode=True)
+
+        # conv2
+        self.conv2_1 = CBR(64, 128, 3, 1, 1)
+        self.conv2_2 = CBR(128, 128, 3, 1, 1)
+        self.pool2 = nn.MaxPool2d(2, stride=2, ceil_mode=True)
+
+        # conv3
+        self.conv3_1 = CBR(128, 256, 3, 1, 1)
+        self.conv3_2 = CBR(256, 256, 3, 1, 1)
+        self.conv3_3 = CBR(256, 256, 3, 1, 1)
+        self.pool3 = nn.MaxPool2d(2, stride=2, ceil_mode=True)
+
+        # conv4
+        self.conv4_1 = CBR(256, 512, 3, 1, 1)
+        self.conv4_2 = CBR(512, 512, 3, 1, 1)
+        self.conv4_3 = CBR(512, 512, 3, 1, 1)
+        self.pool4 = nn.MaxPool2d(2, stride=2, ceil_mode=True)
+
+        # conv5
+        self.conv5_1 = CBR(512, 512, 3, 1, 1)
+        self.conv5_2 = CBR(512, 512, 3, 1, 1)
+        self.conv5_3 = CBR(512, 512, 3, 1, 1)
+        self.pool5 = nn.MaxPool2d(2, stride=2, ceil_mode=True)
+
+        # fc6
+        self.fc6 = CBR(512, 4096, 1, 1, 0)
+        self.drop6 = nn.Dropout2d()
+
+        # fc7
+        self.fc7 = CBR(4096, 4096, 1, 1, 0)
+        self.drop7 = nn.Dropout2d()
+
+        self.score = nn.Conv2d(4096, num_classes, 1, 1, 0)
+
+        self.upscore32 = nn.ConvTranspose2d(
+            num_classes, num_classes, kernel_size=64, stride=32, padding=16)
+
+    def forward(self, x):
+        x = self.conv1_1(x)
+        x = self.conv1_2(x)
+        x = self.pool1(x)
+
+        # conv2
+        x = self.conv2_1(x)
+        x = self.conv2_2(x)
+        x = self.pool2(x)
+
+        # conv3
+        x = self.conv3_1(x)
+        x = self.conv3_2(x)
+        x = self.conv3_3(x)
+        x = self.pool3(x)
+
+        # conv4
+        x = self.conv4_1(x)
+        x = self.conv4_2(x)
+        x = self.conv4_3(x)
+        x = self.pool4(x)
+
+        # conv5
+        x = self.conv5_1(x)
+        x = self.conv5_2(x)
+        x = self.conv5_3(x)
+        x = self.pool5(x)
+
+        x = self.fc6(x)
+        x = self.drop6(x)
+
+        x = self.fc7(x)
+        x = self.drop7(x)
+
+        x = self.score(x)
+        output = self.upscore32(x)
+
+        return output
