+++ junbae/submit.py
@@ -39,5 +39,5 @@ desc = "description"   # 변경 : 파일에 대한 설명
 output_file = "Baseline_FCN8s(pretrained).csv"  # 변경 : output 파일
 
 
-submit(config.submitkey, "./submission/efficientnet_baseline_python_util.csv",
-       "efficient unet baseline 실행 python으로 실행, util수정 miou기준")
+submit(config.submitkey, "./submission/deepv3_vgg16_b8_e20.csv",
+       "deepv3 vgg16(pretrained) b8 e20 테스트 실행")
